// src/lib/database.types.ts
export type Database = {
  public: {
    Tables: {
// ...
      products: {
        Row: {
          category: string | null
          created_at: string | null
          description: string | null
          featured: boolean | null
          heat_level: string | null
          id: string
          image_back_url: string | null
          image_url: string | null
          in_stock: boolean | null
          long_description: string | null
          name: string
          price: number
          slug: string
          sort_order: number | null
          updated_at: string | null
        }
        Insert: {
          category?: string | null
          created_at?: string | null
          description?: string | null
          featured?: boolean | null
          heat_level?: string | null
          id?: string
          image_back_url?: string | null
          image_url?: string | null
          in_stock?: boolean | null
          long_description?: string | null
          name: string
          price: number
          slug: string
          sort_order?: number | null
          updated_at?: string | null
        }
        Update: {
          category?: string | null
          created_at?: string | null
          description?: string | null
          featured?: boolean | null
          heat_level?: string | null
          id?: string
          image_back_url?: string | null
          image_url?: string | null
          in_stock?: boolean | null
          long_description?: string | null
          name?: string
          price?: number
          slug?: string
          sort_order?: number | null
          updated_at?: string | null
        }
        Relationships: []
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}

export type Tables<T extends keyof Database['public']['Tables']> = Database['public']['Tables'][T]['Row']